fix(trusty-code): close cadence env-bleed hole in executor_tests
Extended the fix from cadence_tests.rs to fully close the env-var race
condition by hoisting with_cadence_env and CADENCE_ENV_LOCK into a shared
crate-internal test helper module (cadence_env_helper).

Both cadence_tests.rs::resolve_cadence_config_env_wins_over_settings_json
and executor_tests.rs::settings_json_cadence_turns_override_reaches_resolver
now serialize against the same static lock, ensuring no cargo parallel
scheduler can bleed TCODE_CADENCE_TURNS=5 from one test into another.

The flake class is now fully closed:
- cadence_env_helper: Shared test helper with CADENCE_ENV_LOCK and with_cadence_env
- cadence_tests.rs: Migrated to use shared helper via with_cadence_env import
- executor_tests.rs: Converted to async #[tokio::test], now holds lock via with_cadence_env
- mod.rs: Added cadence_env_helper module (test-only) and re-export

Deterministic repro now passes: TCODE_CADENCE_TURNS=5 cargo test ... --exact settings_json_cadence_turns_override_reaches_resolver
